+++ config/routes.rb
@@ -424,6 +424,7 @@ Rails.application.routes.draw do
         put "update_to_default_docs"
         put "set_colors"
         post "regenerate_images"
+        post "edit_images"
         patch "make_editable"
       end
     end